add delete buttons to meals page
