+++ index.html
@@ -5,6 +5,7 @@
   <title>Quantified Self - Diary</title>
   <link rel="stylesheet" type="text/css" href="lib/diary.css">
   <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
+  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css" integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU" crossorigin="anonymous">
 </head>
 <body>
   <nav class="navbar navbar-expand-lg navbar-light bg-light">
@@ -21,32 +22,40 @@
     </div>
   </nav>
     <h1 class="text-center">Main Diary</h1>
+    <article>
+      <div class="dropdown text-center">
+          <button class="btn btn-info dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+            Add a Food
+          </button>
+          <div class="dropdown-menu" aria-labelledby="dropdownMenuButton"></div>
+        </div>
+    </article>
     <article class="text-center">
       <section class="table-container">
-        <header>Breakfast</header>
+        <button type="button" id="breakfast-btn" class="btn btn-info btn-lg header">Breakfast</button>
         <table class="table table-bordered">
-          <tbody class="Breakfast-table">
+          <tbody class="Breakfast-table meal-table">
           </tbody>
         </table>
       </section>
       <section class="table-container">
-        <header>Lunch</header>
+        <button type="button" id="lunch-btn" class="btn btn-info btn-lg header">Lunch</button>
         <table class="table table-bordered">
-          <tbody class="Lunch-table">
+          <tbody class="Lunch-table meal-table">
           </tbody>
         </table>
       </section>
       <section class="table-container">
-        <header>Dinner</header>
+        <button type="button" id="dinner-btn" class="btn btn-info btn-lg header">Dinner</button>
         <table class="table table-bordered">
-          <tbody class="Dinner-table">
+          <tbody class="Dinner-table meal-table">
           </tbody>
         </table>
       </section>
       <section class="table-container">
-        <header>Snacks</header>
+        <button type="button" id="snack-btn" class="btn btn-info btn-lg header">Snacks</button>
         <table class="table table-bordered">
-          <tbody class="Snack-table">
+          <tbody class="Snack-table meal-table">
           </tbody>
         </table>
       </section>
@@ -54,19 +63,18 @@
     <div class="total-counts">
       <div>
         <h3 class="font-weight-bold d-inline-flex">Total Consumed: </h3>
-        <h3 id="total-calories-consumed" class="d-inline-flex"></h3>
+        <h3 id="total-calories-consumed" class="font-weight-bold d-inline-flex"></h3>
       </div>
       <div>
         <h3 class="font-weight-bold d-inline-flex">Total Goal: </h3>
-        <h3 id="total-goal-calories" class="d-inline-flex"></h3>
+        <h3 id="total-goal-calories" class="font-weight-bold d-inline-flex"></h3>
       </div>
       <div>
         <h3 class="font-weight-bold d-inline-flex">Total Remaining: </h3>
-        <h3 id="total-calories-remaining" class="d-inline-flex"></h3>
+        <h3 id="total-calories-remaining" class="font-weight-bold d-inline-flex"></h3>
       </div>
     </div>
-    <a href="foods.html"><button class="default-button">Create New</button></a>
-
+    <div id="selected-food-id" class="hidden"></div>
   <script
     src='https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js'>
   </script>
