+++ lib/diary.css
@@ -1,7 +1,8 @@
-header {
+.header {
   font-weight: bold;
   text-align: center;
   font-size: 20px;
+  margin: 15px;
 }
 
 .table-container {
@@ -14,4 +15,8 @@ header {
   font-weight: bold;
   text-align: center;
   font-size: 15px;
+}
+
+.hidden {
+  display: none;
 }
