+++ lib/diary.js
@@ -1,0 +1,229 @@
+const getDiaryFoods = () => {
+  fetch("https://fast-meadow-36413.herokuapp.com/api/v1/meals")
+    .then(response => response.json())
+    .then(meals => getMeals(meals))
+    .catch(error => console.error({error}))
+};
+
+const getMeals = (meals) => {
+  meals.forEach(meal => {
+    getMealFoods(meal)
+  });
+  displayTotals();
+  return meals
+};
+
+const getMealFoods = (meal) => {
+  $(`.${meal.name}-table`).html('');
+  meal.foods.forEach(food => {
+    $(`.${meal.name}-table`).append(
+      `<tr>
+        <td class="food-name">${food.name}</td>
+        <td class="food-calories">${food.calories} Cal</td>
+        <td class="trash-square">
+          <a class="btn btn-sm trash-btn" id="${meal.id} ${food.id}" aria-label="Delete">
+            <i class="far fa-trash-alt" aria-hidden="true"></i>
+          </a>
+        </td>
+      </tr>`
+    )
+  });
+  addCaloriesConsumed(meal.name, meal.foods);
+  addGoalCalories(meal.name);
+  addRemainingCalories(meal.name, meal.foods);
+  // <td class="food-id hidden">${food.id}</td>
+};
+
+const addCaloriesConsumed = (mealName, mealFoods) => {
+  const totalCaloriesConsumed = calculateCaloriesConsumed(mealName, mealFoods);
+  $(`.${mealName}-table`).append(
+    `<tr>
+      <td class="font-weight-bold">Consumed</td>
+      <td class="total-meal-calories font-weight-bold">${totalCaloriesConsumed} Cal</td>
+    </tr>`
+  )
+};
+
+const calculateCaloriesConsumed = (mealName, mealFoods) => {
+  const foodCalorieList = Array.from(mealFoods, food => food.calories);
+  return foodCalorieList.reduce((acc, calories) => {
+    return acc + calories
+  }, 0)
+};
+
+const addGoalCalories = (mealName) => {
+  $(`.${mealName}-table`).append(
+    `<tr>
+      <td class="font-weight-bold">Goal</td>
+      <td class="goal-meal-calories font-weight-bold">${findCalorieGoal(mealName)} Cal</td>
+    </tr>`
+  )
+};
+
+const findCalorieGoal = (mealName) => {
+  if (mealName === `Breakfast`) {
+    return 400
+  } else if (mealName === `Lunch`) {
+    return 600
+  } else if (mealName === `Dinner`) {
+    return 800
+  } else if (mealName === `Snack`) {
+    return 200
+  }
+};
+
+const addRemainingCalories = (mealName, mealFoods) => {
+  $(`.${mealName}-table`).append(
+    `<tr>
+      <td class="font-weight-bold">Remaining</td>
+      <td class="remaining-meal-calories font-weight-bold">${findRemainingCalories(mealName, mealFoods)} Cal</td>
+    </tr>`
+  )
+};
+
+const findRemainingCalories = (mealName, mealFoods) => {
+  if (mealName === `Breakfast`) {
+    return (400 - calculateCaloriesConsumed(mealName, mealFoods))
+  } else if (mealName === `Lunch`) {
+    return (600 - calculateCaloriesConsumed(mealName, mealFoods))
+  } else if (mealName === `Dinner`) {
+    return (800 - calculateCaloriesConsumed(mealName, mealFoods))
+  } else if (mealName === `Snack`) {
+    return (200 - calculateCaloriesConsumed(mealName, mealFoods))
+  }
+};
+
+const displayTotals = () => {
+  const totalConsumed = findTotalCaloriesConsumed();
+  const totalGoal = findTotalGoalCalories();
+  const totalRemaining = totalGoal - totalConsumed;
+  $("#total-calories-consumed").html(totalConsumed + ` Cal`);
+  $("#total-goal-calories").html(totalGoal + ` Cal`);
+  $("#total-calories-remaining").html(totalRemaining + ` Cal`);
+};
+
+const findTotalCaloriesConsumed = () => {
+  let totalCalories = 0;
+  $(`.total-meal-calories`).each(function() {
+    totalCalories += parseInt($(this).text().slice(0,-4));
+  });
+  return totalCalories
+};
+
+const findTotalGoalCalories = () => {
+  let totalCalories = 0;
+  $(`.goal-meal-calories`).each(function() {
+    totalCalories += parseInt($(this).text().slice(0,-4));
+  });
+  return totalCalories
+};
+
+const getFoodsForDropDown = () => {
+  fetch("https://fast-meadow-36413.herokuapp.com/api/v1/foods")
+    .then(response => response.json())
+    .then(foods => populateDropDown(foods))
+    .catch(error => console.error({error}))
+};
+
+const populateDropDown = (foods) => {
+  foods.forEach(food => {
+    $(`.dropdown-menu`).append(
+      `<a id="${food.id}" class="dropdown-item">${food.name}</a>`
+    )
+  })
+};
+
+const addBreakfastFood = (food) => {
+  fetch(`https://fast-meadow-36413.herokuapp.com/api/v1/meals/1/foods/${$("#selected-food-id").text()}`, {
+    method: "POST",
+    headers: {"Content-Type": "application/json; charset=utf-8"}
+  })
+  .then(response => response.json())
+  .then(status => checkMealFoodPostStatus(status))
+  .catch(error => console.error({error}))
+};
+
+const addLunchFood = (food) => {
+  fetch(`https://fast-meadow-36413.herokuapp.com/api/v1/meals/3/foods/${$("#selected-food-id").text()}`, {
+    method: "POST",
+    headers: {"Content-Type": "application/json; charset=utf-8"}
+  })
+  .then(response => response.json())
+  .then(status => checkMealFoodPostStatus(status))
+  .catch(error => console.error({error}))
+};
+
+const addDinnerFood = (food) => {
+  fetch(`https://fast-meadow-36413.herokuapp.com/api/v1/meals/4/foods/${$("#selected-food-id").text()}`, {
+    method: "POST",
+    headers: {"Content-Type": "application/json; charset=utf-8"}
+  })
+  .then(response => response.json())
+  .then(status => checkMealFoodPostStatus(status))
+  .catch(error => console.error({error}))
+};
+
+const addSnackFood = (food) => {
+  fetch(`https://fast-meadow-36413.herokuapp.com/api/v1/meals/2/foods/${$("#selected-food-id").text()}`, {
+    method: "POST",
+    headers: {"Content-Type": "application/json; charset=utf-8"}
+  })
+  .then(response => response.json())
+  .then(status => checkMealFoodPostStatus(status))
+  .catch(error => console.error({error}))
+};
+
+const checkMealFoodPostStatus = (status) => {
+  if (status.message.includes("Successfully")) {
+    getDiaryFoods();
+    $(`.dropdown-toggle`).html('Add a Food')
+  } else {
+    $(`.dropdown-toggle`).html('Add a Error')
+  }
+};
+
+$(`.dropdown-menu, .dropdown-item`).click(function(event) {
+  let selected = $(event.target).text();
+  $(`.dropdown-toggle`).html(selected);
+  $(`#selected-food-id`).html(event.target.id)
+});
+
+$(`#breakfast-btn`).click(function(event) {
+  let selectedFood = $(`.dropdown-toggle`).text()
+  if (selectedFood != `Add a Food`) {
+    addBreakfastFood(selectedFood)
+  }
+});
+
+$(`#lunch-btn`).click(function(event) {
+  let selectedFood = $(`.dropdown-toggle`).text()
+  if (selectedFood != `Add a Food`) {
+    addLunchFood(selectedFood)
+  }
+});
+
+$(`#dinner-btn`).click(function(event) {
+  let selectedFood = $(`.dropdown-toggle`).text()
+  if (selectedFood != `Add a Food`) {
+    addDinnerFood(selectedFood)
+  }
+});
+
+$(`#snack-btn`).click(function(event) {
+  let selectedFood = $(`.dropdown-toggle`).text()
+  if (selectedFood != `Add a Food`) {
+    addSnackFood(selectedFood)
+  }
+});
+
+$(`.meal-table, .trash-btn`).click(function(event) {
+  let ids = event.target.id.split(' ');
+  let mealId = ids[0];
+  let foodId = event.target.id[1];
+  
+});
+
+module.exports = {
+  getDiaryFoods: getDiaryFoods,
+  getFoodsForDropDown: getFoodsForDropDown
+};
